Simplify form usage
